requirements.txt
babel==2.17.0
backports.tarfile==1.2.0
boto3==1.36.16
botocore==1.36.16
cached-property==1.5.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
click-spinner==0.1.10
collate-data-diff==0.11.3
collate-sqllineage==1.6.5
colorama==0.4.6
cryptography==44.0.0
daff==1.3.46
dbt-adapters==1.7.0
dbt-artifacts-parser==0.8.2
dbt-common==1.10.0
dbt-core==1.8.8
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
Deprecated==1.2.18
diff_cover==9.2.0
dnspython==2.7.0
docker==7.1.0
dsnparse==0.1.15
email_validator==2.2.0
exceptiongroup==1.2.2
expandvars==0.12.0
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.19.0
google-cloud-storage==1.43.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.63.0rc1
humanfriendly==10.0
idna==3.10
ijson==3.3.0
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.10
mdurl==0.1.2
memory-profiler==0.61.0
minimal-snowplow-tracker==0.0.2
mixpanel==4.10.1
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==2.0.2
openmetadata-ingestion==1.6.3.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pluggy==1.5.0
portalocker==2.10.1
progressbar2==4.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python-utils==3.9.1
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.2
sentry-sdk==2.20.0
shapely==2.0.7
six==1.17.0
SQLAlchemy==1.4.54
sqlfluff==3.1.1
sqlparse==0.5.0
tabulate==0.9.0
tblib==3.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
